refactor: galilean as separate files
